[dao.123] 방탈출 예약 대기 관리 3단계 (#60)
* feat: reservation status 추가

* feat: 예약 승인 기능 구현

* feat: 사용자 예약 취소 기능 구현

* feat: 관리자 예약 취소 기능 구현

* feat: 예약 취소 승인 기능 구현

* feat: Sales Table 추가

* feat: Sales Domain 추가

* fix: Theme Price int -> Long 으로 변경

* feat: 총 매출 조회 기능

* feat: 예약 상태 변경 -> 매출 증감 기능

* feat: 멤버별 매출 조회 기능

* feat: member username unique 제약조건 추가

+++ src/main/java/roomwaiting/nextstep/theme/ThemeDao.java
@@ -32,7 +32,7 @@ public class ThemeDao {
             PreparedStatement ps = connection.prepareStatement(sql, new String[]{"id"});
             ps.setString(1, theme.getName());
             ps.setString(2, theme.getDescription());
-            ps.setInt(3, theme.getPrice());
+            ps.setLong(3, theme.getPrice());
             return ps;
 
         }, keyHolder);
